+++ lib/td_bg/business_concepts/download.ex
@@ -3,6 +3,8 @@ defmodule TdBg.BusinessConcept.Download do
   Helper module to download business concepts.
   """
 
+  alias TdCache.DomainCache
+  alias TdCache.HierarchyCache
   alias TdCache.TemplateCache
   alias TdDfLib.Format
   alias TdDfLib.Templates
@@ -12,6 +14,7 @@ defmodule TdBg.BusinessConcept.Download do
     types = Map.keys(concepts_by_type)
 
     templates_by_type = Enum.reduce(types, %{}, &Map.put(&2, &1, TemplateCache.get_by_name!(&1)))
+    {:ok, domain_name_map} = DomainCache.id_to_name_map()
 
     list =
       Enum.reduce(types, [], fn type, acc ->
@@ -22,7 +25,15 @@ defmodule TdBg.BusinessConcept.Download do
           |> Map.get(type)
           |> Enum.map(&add_completeness(template, &1))
 
-        csv_list = template_concepts_to_csv(template, concepts, header_labels, !Enum.empty?(acc))
+        csv_list =
+          template_concepts_to_csv(
+            template,
+            concepts,
+            header_labels,
+            domain_name_map,
+            !Enum.empty?(acc)
+          )
+
         acc ++ csv_list
       end)
 
@@ -34,23 +45,29 @@ defmodule TdBg.BusinessConcept.Download do
 
   defp add_completeness(_, bcv), do: bcv
 
-  defp template_concepts_to_csv(nil, concepts, header_labels, add_separation) do
+  defp template_concepts_to_csv(nil, concepts, header_labels, domain_name_map, add_separation) do
     headers = build_headers(header_labels)
-    concepts_list = concepts_to_list(concepts)
+    concepts_list = concepts_to_list(concepts, domain_name_map)
     export_to_csv(headers, concepts_list, add_separation)
   end
 
-  defp template_concepts_to_csv(template, concepts, header_labels, add_separation) do
+  defp template_concepts_to_csv(
+         template,
+         concepts,
+         header_labels,
+         domain_name_map,
+         add_separation
+       ) do
     content = Format.flatten_content_fields(template.content)
     content_fields = Enum.reduce(content, [], &(&2 ++ [Map.take(&1, ["name", "values", "type"])]))
     content_labels = Enum.reduce(content, [], &(&2 ++ [Map.get(&1, "label")]))
     headers = build_headers(header_labels)
     headers = headers ++ content_labels
-    concepts_list = concepts_to_list(concepts, content_fields)
+    concepts_list = concepts_to_list(concepts, domain_name_map, content_fields)
     export_to_csv(headers, concepts_list, add_separation)
   end
 
-  defp concepts_to_list(concepts, content_fields \\ []) do
+  defp concepts_to_list(concepts, domain_name_map, content_fields \\ []) do
     Enum.reduce(concepts, [], fn concept, acc ->
       content = concept["content"]
 
@@ -65,7 +82,14 @@ defmodule TdBg.BusinessConcept.Download do
         concept["last_change_at"]
       ]
 
-      acc ++ [Enum.reduce(content_fields, values, &(&2 ++ [&1 |> get_content_field(content)]))]
+      acc ++
+        [
+          Enum.reduce(
+            content_fields,
+            values,
+            &(&2 ++ [&1 |> get_content_field(content, domain_name_map)])
+          )
+        ]
     end)
   end
 
@@ -102,7 +126,7 @@ defmodule TdBg.BusinessConcept.Download do
   defp get_url_value(%{"url_value" => url_value}), do: url_value
   defp get_url_value(_), do: nil
 
-  defp get_content_field(%{"type" => "url", "name" => name}, content) do
+  defp get_content_field(%{"type" => "url", "name" => name}, content, _domain_name_map) do
     content
     |> Map.get(name, [])
     |> content_to_list()
@@ -111,12 +135,40 @@ defmodule TdBg.BusinessConcept.Download do
     |> Enum.join(", ")
   end
 
-  defp get_content_field(%{"type" => type, "name" => name}, content)
-       when type in ["domain", "system"] do
+  defp get_content_field(
+         %{"type" => "hierarchy", "name" => name, "values" => %{"hierarchy" => hierarchy_id}},
+         content,
+         _domain_name_map
+       ) do
+    {:ok, nodes} = HierarchyCache.get(hierarchy_id, :nodes)
+
+    content
+    |> Map.get(name, [])
+    |> content_to_list()
+    |> Enum.map(
+      &Enum.find(nodes, fn %{"node_id" => node_id} ->
+        [_hierarchy_id, content_node_id] = String.split(&1, "_")
+        node_id === String.to_integer(content_node_id)
+      end)
+    )
+    |> Enum.reject(&is_nil/1)
+    |> Enum.map_join(", ", fn %{"name" => name} -> name end)
+  end
+
+  defp get_content_field(%{"type" => "system", "name" => name}, content, _domain_map) do
     content
     |> Map.get(name, [])
     |> content_to_list()
     |> Enum.map(&Map.get(&1, "name"))
+    |> Enum.reject(&is_nil/1)
+    |> Enum.join(", ")
+  end
+
+  defp get_content_field(%{"type" => "domain", "name" => name}, content, domain_name_map) do
+    content
+    |> Map.get(name)
+    |> List.wrap()
+    |> Enum.map(&Map.get(domain_name_map, &1))
     |> Enum.reject(&is_nil/1)
     |> Enum.join(", ")
   end
@@ -127,7 +179,8 @@ defmodule TdBg.BusinessConcept.Download do
            "name" => name,
            "values" => %{"fixed_tuple" => values}
          },
-         content
+         content,
+         _domain_name_map
        ) do
     content
     |> Map.get(name, [])
@@ -139,9 +192,9 @@ defmodule TdBg.BusinessConcept.Download do
     |> Enum.map_join(", ", &Map.get(&1, "text", ""))
   end
 
-  defp get_content_field(%{"type" => "table"}, _content), do: ""
+  defp get_content_field(%{"type" => "table"}, _content, _domain_name_map), do: ""
 
-  defp get_content_field(%{"name" => name}, content) do
+  defp get_content_field(%{"name" => name}, content, _domain_name_map) do
     Map.get(content, name, "")
   end